AdLib & SBPro2 Devices (#190)
* init

* add adlib MidiDriver

* add Adlib device

* [WIP] AdLib device

* [WIP] Adlib midi driver

* [WIP][skip ci] reworking adlib midi driver

* [WIP][skip ci] reworking adlib midi driver

* [WIP][skip ci] reworking adlib midi driver

* [WIP][skip ci] reworking adlib midi driver

* OPL2instrument refactor

* reworking adlib midi driver. first noises

* code rev

* [WIP]reworking adlib midi driver

* [WIP]reworking adlib midi driver

* [WIP] adlib midi driver op2 making some noise

* utils opl

* [WIP] adlib midi driver op2 making some partial music

* adlib device, simplified midi driver

* code rev

* [WIP] pitch bend fix

* pitch bend fix

* pitch bend fixed

* code rev

* sonar cloud rev

* adlib bank op2  midi adlib drv  refactor (#191)

* refactor OP2File, extract audio::opl::banks:Op2Bank_t Op2BankInstrument_t

* replace op2file with audio op2bank

* [WIP] Op2Bank_t -> OP2Bank

* OP2Bank class

* [WIP] refactor

* [wip]

* enable opl callback

* code rev

* utils opl test suite

* fix linux build

* disable test utils opl detectOPL as not working in CI linux & Mac

* std:share_ptr code rev const reference

* code rev

* code rev warnings

* adlib bank op2 midi adlib drv 2nd refactor (#192)

* MUSFile getMidi review

* better debug info

* fix abs_time adlib MidiDriver

* remove MIDIEvent.cpp

* fix build

* remove abs_time in midi event

* fix MIDDriver

* code rev

* code rev

+++ sdl2-sonic-drivers/src/files/VOCFile.cpp
@@ -37,12 +37,12 @@ namespace files
         return _channels;
     }
 
-    const int VOCFile::getSampleRate() const noexcept
+    const uint32_t VOCFile::getSampleRate() const noexcept
     {
         return _sampleRate;
     }
 
-    const int VOCFile::getBitsDepth() const noexcept
+    const uint8_t VOCFile::getBitsDepth() const noexcept
     {
         return _bitsDepth;
     }
